feat: Ex 6.7
Anecdotes, step5.

Moved the Creation form to its own component.
