+++ part06/redux-anecdotes/src/App.js
@@ -1,5 +1,6 @@
 import { useSelector, useDispatch } from 'react-redux';
-import { createAnecdote, voteForAnecdote } from './reducers/anecdoteReducer';
+import AnecdotesForm from './components/AnecdotesForm';
+import { voteForAnecdote } from './reducers/anecdoteReducer';
 
 const App = () => {
   const anecdotes = useSelector((state) =>
@@ -10,13 +11,6 @@ const App = () => {
   const vote = (id) => {
     console.log('vote', id);
     dispatch(voteForAnecdote(id));
-  };
-
-  const create = (event) => {
-    event.preventDefault();
-    const content = event.target.noteContent.value;
-    event.target.noteContent.value = '';
-    dispatch(createAnecdote(content));
   };
 
   return (
@@ -31,13 +25,7 @@ const App = () => {
           </div>
         </div>
       ))}
-      <h2>create new</h2>
-      <form onSubmit={create}>
-        <div>
-          <input name="noteContent" />
-        </div>
-        <button type="submit">create</button>
-      </form>
+      <AnecdotesForm></AnecdotesForm>
     </div>
   );
 };
